Fix up queries.
Need to "left join" media, as not all media fields will be related to
every media entity (really, we only expect one for each). Similarly,
some media might not make use of files, but we still want to track in
the LUT to facilitate media checks, so let's "left join" to files and
null coalesce back to the table's default.